Rename player-state service
